* add simple non-standard repo filtering to `renv_lock_pkgs()` function
---

adds the `exclude_other_sources` option to `renv_lock_pkgs()` and
defaults it to TRUE.
This excludes packages from being returned if the do not have
"Repository" as their value for "Source" in renv.lock the and throws a
warning listing the excluded packages.
This should exclude packages from sources like github which are not
presently like to be packaged in nix.