* add `renv_remote_pkgs()` function, factors out generating `git_pkgs` information
---

The `renv_remote_pkgs()` function takes a single renv.lock package entry
as an argument, infers it's remote type and extracts the informaton
necesary for `rix()`'s `git_pkgs` argument.

This functionality was factored out of the `renv2nix()`, and the
`method` argument added `renv2nix()` to allow for a future
implementation of exact package version matching.